Preprocessing gui design done
The design of the preprocessing gui is done. Now we should start adding
functionalities to the interface. The default format for exported would be json. Supporting TEI, and AltoXML is an option for the future
Some sort of hierarchy between the selected regions is essential.
The interface is to be as simple as possible. An needs to be throughly tested.
Test suite for the interface should be structured.