Started to a new interface with js and html/css
Started on the new interface with js and html

Sections and general layout is done.

Need to implement image selection.
Now we can select images:
We need to see their names in the image list widget
then we need to be able to load the selected one to the
svg viewer on the viewer section