Added new toolset and selector code
Added selector code and the new toolset for the viewer

SVG seems to sets draggable as default for its dom elements need to change
that.

The image is a background image, and all the rest are new layers added upon
the image.

One way of doing this is defining the image as a pattern and then refering to
it in paths but I don't have a lot of hope for it. Instead I should aim for
drawing over the image by making it appear as a background in all the
svg canvas rather than the background of a specific item like path