New methods to draw objects
The idea is pretty simple there are two sides to drawing objects:

1. Draw SVG objects to canvas then draw the object that is being drawn
at the moment.

2. Pass newly drawn object to svg dom.

The passage between svg objects to newly drawn object passes through
the passing of the "context" of drawing,

Adding newly drawn object to svg is a little more difficult.
But globally speaking the adding should occur at mouse up.
The drawing of svg objects and the current object should occur during
mousedown and mousemove.
There should be an option for showing all the svg objects or not.
We can change their visibility.