New approach to drawing started
I decided to dump the svg + canvas approach.
I'll just go with the canvas and make an stack of objects that are drawn.
The mousemove and mousedown triggers drawing.
The method for stacking up the drawn objects to array should be done with a
button. There should be a method for holding several drawn objects at the same
time on the scene. This is again done with a temporary stack.